ytLearn/requirements.txt
youtube-transcript-api
yt-dlp
flask
markdown
google-generativeai
python-dotenv
